+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 4.3-SNAPSHOT
   Bugs
+   * Fix #1667: Origin header for watch requests had a port of -1 when no port specified
 
   Improvements
    * Test coverage for PersistentVolumeClaim